# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/history/apply": {
      "post": {
        "operationId": "apply_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Op"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlobRef": {
        "type": "string",
        "description": "Hex-encoded blake3 hash of an immutable blob."
      },
      "Chapter": {
        "type": "object",
        "required": [
          "id",
          "name",
          "order",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/ChapterId"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageId"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChapterId": {
        "type": "string",
        "format": "uuid"
      },
      "ChapterPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "order": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "pageIds": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/PageId"
            }
          }
        }
      },
...
      "FontPrediction": {
        "type": "object",
        "required": [
          "topFonts",
          "namedFonts",
          "direction",
          "textColor",
          "strokeColor",
          "fontSizePx",
          "strokeWidthPx",
          "lineHeight",
          "angleDeg"
        ],
        "properties": {
          "angleDeg": {
            "type": "number",
            "format": "float"
          },
          "direction": {
            "$ref": "#/components/schemas/TextDirection"
          },
          "fontSizePx": {
            "type": "number",
            "format": "float"
          },
          "lineHeight": {
            "type": "number",
            "format": "float"
          },
          "namedFonts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedFontPrediction"
            }
          },
          "strokeColor": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          "strokeWidthPx": {
            "type": "number",
            "format": "float"
          },
          "textColor": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          "topFonts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopFont"
            }
          }
        }
      },
...
      "HistoryResult": {
        "type": "object",
        "properties": {
          "epoch": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "New epoch. `None` only for a no-op undo/redo at the stack boundary.",
            "minimum": 0
          }
        }
      },
...
      "ImageData": {
        "type": "object",
        "required": [
          "role",
          "blob",
          "naturalWidth",
          "naturalHeight"
        ],
        "properties": {
          "blob": {
            "$ref": "#/components/schemas/BlobRef"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "naturalHeight": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "naturalWidth": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "opacity": {
            "type": "number",
            "format": "float"
          },
          "role": {
            "$ref": "#/components/schemas/ImageRole",
            "description": "Role tags differentiate source / inpainted / rendered / user-imported images.\nRole is immutable on an existing node — switching roles = delete + add."
          }
        }
      },
      "ImageDataPatch": {
        "type": "object",
        "properties": {
          "blob": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BlobRef"
              }
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "naturalHeight": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "naturalWidth": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "opacity": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          }
        }
      },
      "ImageRole": {
        "type": "string",
        "enum": [
          "source",
          "inpainted",
          "rendered",
          "custom"
        ]
      },
...
      "MaskData": {
        "type": "object",
        "required": [
          "role",
          "blob"
        ],
        "properties": {
          "blob": {
            "$ref": "#/components/schemas/BlobRef"
          },
          "role": {
            "$ref": "#/components/schemas/MaskRole"
          }
        }
      },
      "MaskDataPatch": {
        "type": "object",
        "properties": {
          "blob": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BlobRef"
              }
            ]
          }
        }
      },
      "MaskRole": {
        "type": "string",
        "enum": [
          "brushInpaint",
          "segment",
          "bubble"
        ]
      },
...
      "NamedFontPrediction": {
        "type": "object",
        "required": [
          "index",
          "name",
          "probability",
          "serif"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "language": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "probability": {
            "type": "number",
            "format": "float"
          },
          "serif": {
            "type": "boolean"
          }
        }
      },
      "Node": {
        "type": "object",
        "required": [
          "id",
          "visible",
          "kind"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/NodeId"
          },
          "kind": {
            "$ref": "#/components/schemas/NodeKind"
          },
          "transform": {
            "$ref": "#/components/schemas/Transform"
          },
          "visible": {
            "type": "boolean"
          }
        }
      },
      "NodeDataPatch": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "text"
            ],
            "properties": {
              "text": {
                "$ref": "#/components/schemas/TextDataPatch"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "image"
            ],
            "properties": {
              "image": {
                "$ref": "#/components/schemas/ImageDataPatch"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "mask"
            ],
            "properties": {
              "mask": {
                "$ref": "#/components/schemas/MaskDataPatch"
              }
            }
          }
        ]
      },
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
      "NodeKind": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "image"
            ],
            "properties": {
              "image": {
                "$ref": "#/components/schemas/ImageData"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "text"
            ],
            "properties": {
              "text": {
                "$ref": "#/components/schemas/TextData"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "mask"
            ],
            "properties": {
              "mask": {
                "$ref": "#/components/schemas/MaskData"
              }
            }
          }
        ]
      },
      "NodePatch": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NodeDataPatch"
              }
            ]
          },
          "transform": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Transform"
              }
            ]
          },
          "visible": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
      "Op": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "updateProjectMeta"
            ],
            "properties": {
              "updateProjectMeta": {
                "type": "object",
                "required": [
                  "patch"
                ],
                "properties": {
                  "patch": {
                    "$ref": "#/components/schemas/ProjectMetaPatch"
                  },
                  "prev": {
                    "$ref": "#/components/schemas/ProjectMetaPatch"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "addPage"
            ],
            "properties": {
              "addPage": {
                "type": "object",
                "required": [
                  "page",
                  "at"
                ],
                "properties": {
                  "at": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "page": {
                    "$ref": "#/components/schemas/Page"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "removePage"
            ],
            "properties": {
              "removePage": {
                "type": "object",
                "required": [
                  "id",
                  "prev_page",
                  "prev_index"
                ],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "prev_index": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "prev_page": {
                    "$ref": "#/components/schemas/Page"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "updatePage"
            ],
            "properties": {
              "updatePage": {
                "type": "object",
                "required": [
                  "id",
                  "patch"
                ],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "patch": {
                    "$ref": "#/components/schemas/PagePatch"
                  },
                  "prev": {
                    "$ref": "#/components/schemas/PagePatch"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "reorderPages"
            ],
            "properties": {
              "reorderPages": {
                "type": "object",
                "required": [
                  "order",
                  "prev_order"
                ],
                "properties": {
                  "order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PageId"
                    }
                  },
                  "prev_order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PageId"
                    }
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "addChapter"
            ],
            "properties": {
              "addChapter": {
                "type": "object",
                "required": [
                  "chapter"
                ],
                "properties": {
                  "chapter": {
                    "$ref": "#/components/schemas/Chapter"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "removeChapter"
            ],
            "properties": {
              "removeChapter": {
                "type": "object",
                "required": [
                  "id",
                  "prev_chapter",
                  "prev_index"
                ],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/ChapterId"
                  },
                  "prev_chapter": {
                    "$ref": "#/components/schemas/Chapter"
                  },
                  "prev_index": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "updateChapter"
            ],
            "properties": {
              "updateChapter": {
                "type": "object",
                "required": [
                  "id",
                  "patch"
                ],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/ChapterId"
                  },
                  "patch": {
                    "$ref": "#/components/schemas/ChapterPatch"
                  },
                  "prev": {
                    "$ref": "#/components/schemas/ChapterPatch"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "reorderChapters"
            ],
            "properties": {
              "reorderChapters": {
                "type": "object",
                "required": [
                  "order",
                  "prev_order"
                ],
                "properties": {
                  "order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ChapterId"
                    }
                  },
                  "prev_order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ChapterId"
                    }
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "addNode"
            ],
            "properties": {
              "addNode": {
                "type": "object",
                "required": [
                  "page",
                  "node",
                  "at"
                ],
                "properties": {
                  "at": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "node": {
                    "$ref": "#/components/schemas/Node"
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "removeNode"
            ],
            "properties": {
              "removeNode": {
                "type": "object",
                "required": [
                  "page",
                  "id",
                  "prev_node",
                  "prev_index"
                ],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/NodeId"
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "prev_index": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "prev_node": {
                    "$ref": "#/components/schemas/Node"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "updateNode"
            ],
            "properties": {
              "updateNode": {
                "type": "object",
                "required": [
                  "page",
                  "id",
                  "patch"
                ],
                "properties": {
                  "id": {
                    "$ref": "#/components/schemas/NodeId"
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "patch": {
                    "$ref": "#/components/schemas/NodePatch"
                  },
                  "prev": {
                    "$ref": "#/components/schemas/NodePatch"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "reorderNodes"
            ],
            "properties": {
              "reorderNodes": {
                "type": "object",
                "required": [
                  "page",
                  "order",
                  "prev_order"
                ],
                "properties": {
                  "order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/NodeId"
                    }
                  },
                  "page": {
                    "$ref": "#/components/schemas/PageId"
                  },
                  "prev_order": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/NodeId"
                    }
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "batch"
            ],
            "properties": {
              "batch": {
                "type": "object",
                "required": [
                  "ops",
                  "label"
                ],
                "properties": {
                  "label": {
                    "type": "string"
                  },
                  "ops": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Op"
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
      "Page": {
        "type": "object",
        "required": [
          "id",
          "name",
          "width",
          "height",
          "nodes"
        ],
        "properties": {
          "chapterId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChapterId"
              }
            ]
          },
          "height": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "id": {
            "$ref": "#/components/schemas/PageId"
          },
          "name": {
            "type": "string"
          },
          "nodes": {
            "type": "object",
            "description": "Stacking = insertion order. Bottom-first: `source` is typically first,\n`rendered` typically last.",
            "additionalProperties": {
              "$ref": "#/components/schemas/Node"
            },
            "propertyNames": {
              "type": "string",
              "format": "uuid"
            }
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
      "PagePatch": {
        "type": "object",
        "properties": {
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ProjectMetaPatch": {
        "type": "object",
        "properties": {
          "mangaId": {
            "type": [
              "string",
              "null"
            ]
          },
          "mangaTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "sourceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "style": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProjectStyle"
              }
            ]
          },
          "syncDir": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "ProjectStyle": {
        "type": "object",
        "properties": {
          "defaultFont": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TextData": {
        "type": "object",
        "properties": {
          "confidence": {
            "type": "number",
            "format": "float"
          },
          "detectedFontSizePx": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "detector": {
            "type": [
              "string",
              "null"
            ]
          },
          "fontPrediction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FontPrediction"
              }
            ]
          },
          "linePolygons": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          },
          "lockLayoutBox": {
            "type": "boolean"
          },
          "renderedDirection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextDirection"
              }
            ]
          },
          "rotationDeg": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "sourceDirection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextDirection"
              }
            ]
          },
          "sourceLang": {
            "type": [
              "string",
              "null"
            ]
          },
          "sprite": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BlobRef",
                "description": "Renderer-produced sprite for this block."
              }
            ]
          },
          "spriteTransform": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Transform",
                "description": "Sprite placement when the renderer expands past the bubble geometry."
              }
            ]
          },
          "style": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextStyle"
              }
            ]
          },
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "translation": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TextDataPatch": {
        "type": "object",
        "description": "For fields where \"set to None\" is meaningful (e.g. clearing a translation),\nthe outer `Option` is \"patch present\", the inner is \"value present\".",
        "properties": {
          "confidence": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "detectedFontSizePx": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "detector": {
            "type": [
              "string",
              "null"
            ]
          },
          "fontPrediction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FontPrediction"
              }
            ]
          },
          "linePolygons": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          },
          "lockLayoutBox": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "renderedDirection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextDirection"
              }
            ]
          },
          "rotationDeg": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "sourceDirection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextDirection"
              }
            ]
          },
          "sourceLang": {
            "type": [
              "string",
              "null"
            ]
          },
          "sprite": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BlobRef"
              }
            ]
          },
          "spriteTransform": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Transform"
              }
            ]
          },
          "style": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextStyle"
              }
            ]
          },
          "text": {
            "type": [
              "string",
              "null"
            ]
          },
          "translation": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TextDirection": {
        "type": "string",
        "description": "Reading axis of a text block.",
        "enum": [
          "horizontal",
          "vertical"
        ]
      },
...
      "TopFont": {
        "type": "object",
        "required": [
          "index",
          "score"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "score": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "Transform": {
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "properties": {
          "height": {
            "type": "number",
            "format": "float"
          },
          "rotationDeg": {
            "type": "number",
            "format": "float"
          },
          "width": {
            "type": "number",
            "format": "float"
          },
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          }
        }
      }
    }
  }
}